Fix label generation and implement complete data processing pipeline
Major changes:
- Fix Label 1 (CIPS) with robust multi-strategy antibody detection
- Implement Label 2 (BepiPred/Ellipro) integration with HPC-aware caching
- Fix ESM2 loading hang issue with lazy loading
- Add compute node detection for offline operation
- Create pre-processing script for BepiPred/Ellipro predictions
- Fix ROC AUC error handling for single-class evaluation
- Standardize on SLURM testing workflow

New files:
- data_processing/epi4ab_pipeline.py: Complete end-to-end pipeline (1,489 lines)
- preprocess_epitope_predictions.py: Pre-fetch script for predictions
- test_label_generation.py: Diagnostic test script
- Comprehensive documentation in docs/

Modified files:
- interface_prediction/model/testing_function.py: ROC AUC fix
- interface_prediction/data_function/data_function.py: Various fixes
- interface_prediction/run_setup/arguments.py: Updates

+++ interface_prediction/run_setup/arguments.py
@@ -71,7 +71,7 @@ def initiate_argument():
     #                     action='store_true', default=True)
     parser.add_argument('--bias_distance', metavar='',
                         help='Value for substracting distance of attribute',
-                        default=0, type=int)
+                        default=0, type=float)
     parser.add_argument('--softmax_data',
                         help='Set softmax layer for data before training',
                         action='store_true')
